requirements.txt
alabaster==0.7.12
asttokens==2.2.1
attr==0.3.2
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
boltons==21.0.0
branca==0.5.0
Brotli==1.0.9
brotlicffi==1.0.9.2
bs4==0.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.1.2
configparser==5.3.0
contourpy==1.0.6
cryptography==38.0.1
cs-brix==0.0.10
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.5
decorator==5.1.1
dl==0.1.0
docutils==0.19
entrypoints==0.4
executing==1.2.0
folium==0.13.0
fonttools==4.38.0
funcsigs==1.0.2
geographiclib==2.0
geopandas==0.10.2
geopy==2.3.0
h3==3.7.6
imagesize==1.4.1
importlib-metadata==5.0.0
ipykernel==6.20.1
ipython==8.5.0
ipywidgets==8.0.2
jaraco.classes==3.2.3
jedi==0.18.2
jellyfish==0.6.1
Jinja2==3.1.2
jupyter_client==7.4.8
jupyter_core==5.1.3
jupyterlab-widgets==3.0.5
keyring==23.9.3
kiwisolver==1.4.4
lockfile==0.12.2
lxml==4.9.2
mapclassify==2.4.3
MarkupSafe==2.1.1
matplotlib==3.6.1
more-itertools==9.0.0
munch==2.5.0
nest-asyncio==1.5.6
networkx==3.0
numpy==1.23.4
opencity==1.0.1
ordereddict==1.1
osmnet==0.1.6
osmnx==1.2.0
packaging==23.0
pandana==0.6.1
pandas==1.5.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
protobuf==4.21.7
psutil==5.9.4
pure-eval==0.2.2
pygeohash==1.2.0
Pygments==2.14.0
pyOpenSSL==22.1.0
#pyproj==3.4.1
scipy==1.9.2
Shapely==1.8.5.post1
soupsieve==2.3.2.post1
statsmodels==0.13.2
tables==3.6.1
tabulate==0.9.0
us==2.0.2

